reorganized top level processes
`tlp`, or `top level process`, is a term from occam. It means a PROC
that is defined at the top level, or otherwise is exposed to the
namespace. Typically, a TLP will run forever, although not strictly
true.

To clean up the main file (`session-counter.go`), I moved all of the
PROCs (gofuncs containing a `for { ... }`, or infinite loop) into a tlp
package. This got rid of the tshark package by putting all the stuff
that belongs together, together.

At the top level,

go test *.go

will run tests on the new algorithm (which was committed yesterday).
